feat(stm32mp1): manage low-power
When returning from Standby mode, BL2 will be in charge of putting the
DDR out of self-refresh mode, and do the correct initialization,
different from a cold boot. This is done thanks to information saved
in backup SRAM in a low-power context structure.
Some loading of images should be skipped, as already in DDR.

If the monitor is in DDR, which is the case on STM32MP13 chip with
OP-TEE, some code in BL2, resident in SYSRAM, will be in charge of
setting the DDR in self-refresh mode.

Change-Id: Ie1a12b24a9833b95a233774c74bb76114c15380d
Signed-off-by: Yann Gautier <yann.gautier@st.com>