feat(stm32mp2): derivating DDRCryp encryption key from HUK
This patch changes how the DDRCryp key is generated. Instead of using
a random key, we use a derivative key from the SAES hardware key. We
just have to store the 32-bit random seed in the backup context instead
of the full key.

Signed-off-by: Maxime Méré <maxime.mere@foss.st.com>
Change-Id: Ibea4423274cf15e6c9482131fa6c2a7eac6ab853
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/353733
Domain-Review: Sebastien PASDELOUP <sebastien.pasdeloup-ext@st.com>
Reviewed-by: Sebastien PASDELOUP <sebastien.pasdeloup-ext@st.com>
Tested-by: Sebastien PASDELOUP <sebastien.pasdeloup-ext@st.com>