feat(stm32mp25-fdts): update RIF configuration
Align the "open" configuration with the latest version of the xls tool.
It fixes the bl31_context area (start addr + size) and allows the CID0 to
access to the "op-tee" area (as it is done for all area for a "open"
configuration).

It also changes PLL binding to satisfy YAML verification and moves U(S)ART
source clock to HSI_KER for low power feature.

Change-Id: I8224568b90eb42a99c90b727df063792be17780d
Signed-off-by: Alexandre TORGUE <alexandre.torgue@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/363565
Domain-Review: Sebastien PASDELOUP <sebastien.pasdeloup-ext@st.com>
Reviewed-by: Sebastien PASDELOUP <sebastien.pasdeloup-ext@st.com>
Tested-by: Sebastien PASDELOUP <sebastien.pasdeloup-ext@st.com>