feat(stm32mp25): adapt authentication status
In the STM32 MPU, we use `auth_status` from `boot_api_context` to
display the authentication check status on the console for CLOSED
UNSECURED/UNLOCKED samples. On the MP15 and MP13, this `auth_status` was
consistent with the behavior of the sample once SECURED/LOCKED. However,
on the MP2, this is no longer the case; it reflects only the Signature
check and not the ROTPK check.

This patch aims to resolve this issue by adding additional checks to
indicate whether the authentication is successful. The code is placed
under the `TRUSTED_BOARD_BOOT` and `DYN_DISABLE_AUTH` precompiler flags,
as it is necessary for debugging purposes and not in a real trusted
mode.

Signed-off-by: Maxime Méré <maxime.mere@foss.st.com>
Change-Id: Ide4f46e97ab6d1db1bb04cb1f627762f7783ce7f
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/375903
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>